+++ changelog.md
@@ -1,5 +1,6 @@
 = 0.4.9 =
 * Fix: Custom variation attributes
+* Fix (WC3 compatibility): non-taxable products
 
 = 0.4.8 - 2017/04/27 =
 * Fix: 'Any ...' variations